requirements.txt
lxml==5.2.1
python-docx==1.1.0
python-dotenv==1.0.1
typing_extensions==4.11.0
